+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "source-engine-support" extension will be documented 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.8]
+
+**VMT**
+- Added color value highlighting and validation
+- Added snippets for material templates
+
 ## [0.0.7]
 
 **VMT**